refactor(ui): split StreamSnackbar into component and messenger parts
Move the orchestration (StreamSnackbarMessenger, StreamSnackbarController,
StreamSnackbarClosedReason, the Host/Popup/Scope surfaces, and the shared
private rendering pipeline) into a `part of` companion file so the
entrypoint focuses on the component itself and its theming.
